+++ src/utils.py
@@ -226,25 +226,4 @@ def get_transactions_for_categories(data: list[dict[str, Any]], categories: set)
     return [tx for tx in data if tx.get(CATEGORY_KEY, "") in categories]
 
 
-def get_cashback_categories(data: list[dict[str, Any]], percent_cashback: float) -> dict[str, Any]:
-    """Получает на вход список транзакций возвращает возможный кэшбек по категориям"""
-    data_cashback = [
-        {**tx, AMOUNT_ROUND_UP_KEY: int(tx.get(AMOUNT_ROUND_UP_KEY, 0.0) * percent_cashback / 100)} for tx in data
-    ]
 
-    result = get_amount_for_categories(data_cashback)
-    logger.info("Получены словари кэшбека по категориям")
-    return result
-
-
-def get_invest_amount(data: list[dict[str, Any]], limit: int) -> float:
-    """Получает на вход список транзакций и лимит округления, возвращает сумму"""
-    lst = [(tx.get(AMOUNT_ROUND_UP_KEY, 0) // limit + 1) * limit - tx.get(AMOUNT_ROUND_UP_KEY, 0) for tx in data]
-    return float(sum(lst))
-
-
-def get_search_by_keyword(data: list[dict[str, Any]], keyword: str, search_keys: set) -> list[dict[str, Any]]:
-    """Получает на вход список транзакций, запрос и множество ключей поиска, возвращает список транзакций"""
-    pattern = re.compile(re.escape(keyword), re.IGNORECASE)
-    result = [tx for tx in data if any(pattern.search(tx[key]) for key in search_keys)]
-    return result
